Complete SoarUnit feature
Use the default `exec` plugin to properly display the output of the
SoarUnit execution, and allow the user to click on the output and
go to the tested file.

Update the readme with directions on using this feature, and for now
tell the user not to use the full project build feature. The recursive
run feature of SoarUnit seems to be very buggy.